LMS-1697 Fixed minor bugs with graphs that popped up as a result of recent work.
SVN: 17574

+++ screening/source/java/ch/systemsx/cisd/openbis/dss/generic/server/graph/TabularDataHeatmap.java
@@ -216,7 +216,7 @@ public class TabularDataHeatmap extends AbstractTabularDataGraph<TabularDataHeat
             {
                 try
                 {
-                    element.x = Integer.parseInt(line[xColumn]);
+                    element.x = (int) Float.parseFloat(line[xColumn]);
                 } catch (NumberFormatException ex)
                 {
                     // handle a case when X is alphanumeric
@@ -224,7 +224,7 @@ public class TabularDataHeatmap extends AbstractTabularDataGraph<TabularDataHeat
                             Location.tryCreateLocationFromTransposedMatrixCoordinate(
                                     line[xColumn] + "1").getY();
                 }
-                element.y = Integer.parseInt(line[yColumn]);
+                element.y = (int) Float.parseFloat(line[yColumn]);
             } else
             {
                 Location loc =
